refactor: make generalized AlertContext and AlertProvider
