+++ src/sections/dataset/DatasetFactory.tsx
@@ -11,7 +11,7 @@ import { SettingJSDataverseRepository } from '../../settings/infrastructure/Sett
 import { FilePermissionsProvider } from '../file/file-permissions/FilePermissionsProvider'
 import { SettingsProvider } from '../settings/SettingsProvider'
 import { DatasetProvider } from './DatasetProvider'
-import { DatasetAlertProvider } from './DatasetAlertProvider'
+import { AlertProvider } from '../alerts/AlertProvider'
 
 const datasetRepository = new DatasetJSDataverseRepository()
 const fileRepository = new FileJSDataverseRepository()
@@ -25,9 +25,9 @@ export class DatasetFactory {
         <SettingsProvider repository={settingRepository}>
           <MetadataBlockInfoProvider repository={metadataBlockInfoRepository}>
             <AnonymizedProvider>
-              <DatasetAlertProvider>
+              <AlertProvider>
                 <DatasetWithSearchParams />
-              </DatasetAlertProvider>
+              </AlertProvider>
             </AnonymizedProvider>
           </MetadataBlockInfoProvider>
         </SettingsProvider>
